Refactor DSV4 Async CAM forward
