+++ afd_plugin/model_executor/models/model_utils.py
@@ -19,9 +19,6 @@ def get_afd_model_config(
     device_type: Literal["cuda", "npu"],
 ) -> ModelConfig:
     """Return a model config that resolves to an AFD model implementation."""
-    from afd_plugin import register_afd
-    register_afd()
-
     for model_arch in model_config.hf_config.architectures:
         if model_arch in _MODEL_REGISTRATIONS:
             if model_arch in _QWEN3_5_MODEL_REGISTRATIONS and device_type != "cuda":
